style: Rename files to match python conventions
