+++ test/test_database.py
@@ -1,11 +1,11 @@
 from unittest.mock import patch
 import asynctest
 import mockredis
-from bowser.Database import Database
+from bowser.database import Database
 
 
 class TestDatabase(asynctest.TestCase):
-    @patch('bowser.Database.redis.StrictRedis', mockredis.mock_strict_redis_client)
+    @patch('bowser.database.redis.StrictRedis', mockredis.mock_strict_redis_client)
     def test__able_to_set_and_fetch_data(self):
         db = Database()
         db.set_data_of_server_channel('fake_server_id', 'fake_channel', {'host': 'some_host', 'port': 1234})
